init But some function is always completed.
DESCRIPTION FILE
[MGG.js]
This file is function of thought of MGG.

[block_break.html]
This file is main programe.
View, basic function of functions that is include moving ball,bar,block and calucurate GA are here.

[choice.js]
This file is function of choice for GA.
Rulet and Elite Method are here.

[crossover.js]
This file is function of crossover for GA.
Single and Double crossover function are here.

[selection.js]
This file is deleting weak individual and adding child.